Refactoring - useClickOutOfBlock hook
