+++ src/components/Layout/Header/Droppdown/Droppdown.tsx
@@ -1,32 +1,16 @@
-import { useEffect, useRef, useState } from "react";
+import { useRef, useState } from "react";
 import { IoChevronDownOutline } from "react-icons/io5";
 
 import { P } from "../../../UI";
 import styles from "./Droppdown.module.css";
 import { Content } from "./Content";
+import { useClickOutOfBlock } from "../../../../hooks/useClickOutOfBlock";
 
 export const Droppdown = (): JSX.Element => {
   const [visible, setVisible] = useState(false);
   const dropdownRef = useRef<HTMLDivElement | null>(null);
 
-  // handling click outside of the droppdown to close it
-
-  useEffect(() => {
-    const handleClickOutside = (event: MouseEvent): void => {
-      if (
-        dropdownRef.current &&
-        !dropdownRef.current.contains(event.target as Node)
-      ) {
-        setVisible(false);
-      }
-    };
-
-    document.addEventListener("click", handleClickOutside);
-
-    return () => {
-      document.removeEventListener("click", handleClickOutside);
-    };
-  }, []);
+  useClickOutOfBlock(dropdownRef, setVisible);
 
   return (
     <div ref={dropdownRef} className={styles.droppdown}>
