+++ src/components/Layout/Header/Home.module.css
@@ -5,7 +5,6 @@
   align-items: center;
   z-index: 3;
   background: var(--white);
-  box-shadow: 0px 0px 10px 0px rgba(0, 0, 0, 0.05);
 }
 
 .logo {
